pyproject.toml
[project]
name = "aegis"
version = "0.1.0"
description = "Security-first isolated execution environment and scanning gateway for AI agents"
requires-python = ">=3.12"
license = { text = "MIT" }
readme = "README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aegis"]

[dependency-groups]
docs = [
    "mkdocs-material>=9.6",
    "mkdocs-mermaid2-plugin>=1.1",
]
dev = [
    "pytest>=8.0",
    "httpx>=0.28",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0",
    "mitmproxy>=11.0",
]
